% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemetadb.R
\docType{data}
\name{ae_arrays}
\alias{ae_arrays}
\title{ArrayExress arrays table}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 735 rows and 3 columns.}
\source{
ArrayExpress @ EBI
}
\usage{
ae_arrays
}
\description{
subset of array express arrays that match:
  'saccharomyces+cerevisiae'
}
\keyword{datasets}

