% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemetadb.R
\docType{data}
\name{ae_experiments}
\alias{ae_experiments}
\title{ArrayExress experiments table}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1566 rows and 10 columns.}
\source{
ArrayExpress @ EBI
}
\usage{
ae_experiments
}
\description{
subset of array express experiments that match:
  'saccharomyces+cerevisiae'
}
\keyword{datasets}

