% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REST-api-experiments.R
\name{ae_get_files}
\alias{ae_get_files}
\alias{ae_get_protocols}
\alias{ae_get_samples}
\title{get files associated with an accession}
\usage{
ae_get_files(accession, tidy = T)

ae_get_protocols(accession, tidy = T)

ae_get_samples(accession, tidy = T)
}
\arguments{
\item{accession}{<chr> Array Express Experiment accession}

\item{tidy}{<lgl> should the response contents be tidied up?}
}
\value{
a tidy tibble by default, httr::GET response object if tidy = F
}
\description{
get files associated with an accession
}

