% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REST-api-experiments.R
\name{ae_query}
\alias{ae_query}
\title{general query against api}
\usage{
ae_query(url = ae_rest_url(), ...)
}
\arguments{
\item{url}{<chr> of url endpoint to query, see ae_rest_url()}

\item{...}{query parameters}
}
\value{
httr::GET response object
}
\description{
general query against api
}
\details{
http://www.ebi.ac.uk/arrayexpress/help/programmatic_access.html
}
\examples{
  ae_query(keywords = "cancer")
  ae_rest_url("files") \%>\% ae_query(array = "A-AFFY-33")
}

