% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REST-api-experiments.R
\name{ae_rest_url}
\alias{ae_rest_url}
\title{Show the url for a given Array Express REST API}
\usage{
ae_rest_url(api = c("experiments", "files", "protocols"),
  rettype = c("json", "xml"), version = "v3")
}
\arguments{
\item{api}{<chr> which api}

\item{rettype}{<chr> which format return type}

\item{version}{<chr> which version of api}
}
\value{
<chr> url of API endpoint
}
\description{
Show the url for a given Array Express REST API
}

