% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae-files.R
\name{ae_sample_files}
\alias{ae_sample_files}
\title{get sample files associated with an experiment}
\usage{
ae_sample_files(experiment)
}
\arguments{
\item{experiment}{array express exeperiment accession as character string.}
}
\description{
get sample files associated with an experiment
}
\examples{
  # ae_sample_files("E-MEXP-27")
}

