% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_in_arrays}
\alias{find_in_arrays}
\title{find arrays by searching for match in whole table}
\usage{
find_in_arrays(x)
}
\arguments{
\item{x}{pattern to be used in stringr::str_detect}
}
\description{
find arrays by searching for match in whole table
}

